rbottini - add test and configmap

+++ src/main/java/com/riccardobottini/chaosmonkey/DeletePodScheduledJob.java
@@ -27,6 +27,9 @@ public class DeletePodScheduledJob {
     @Value("${k8s.target.namespace}")
     private String namespace;
 
+    @Value("${k8s.target.namespace.pod.label}")
+    private String label;
+
     private Random random = new Random();
 
     private static final Logger log = LoggerFactory.getLogger(DeletePodScheduledJob.class);
@@ -39,63 +42,72 @@ public class DeletePodScheduledJob {
 
         ApiClient apiClient;
         try {
+            // Configure K8s API client
             apiClient = Config.defaultClient();
             Configuration.setDefaultApiClient(apiClient);
             CoreV1Api coreV1Api = new CoreV1Api(apiClient);
 
             log.info("Getting pods from {}", namespace);
 
+            // List pods
             V1PodList podList = coreV1Api.listNamespacedPod(
-                namespace, 
-                null, 
-                null, 
-                null, 
-                null, 
-                null, 
-                null, 
-                null, 
-                null,
-                null, 
-                null
-            );
-
-            for (V1Pod pod : podList.getItems()){
-                log.info("Pod name: {}", pod.getMetadata().getName());
-            }
-            
-            int randomListElementIndex = random.nextInt(podList.getItems().size());
-
-            V1Pod podToBeDeleted = podList.getItems().get(randomListElementIndex);
-            V1ObjectMeta podMetadata = podToBeDeleted.getMetadata();
-
-            if (podMetadata != null) {
-                V1DeleteOptions v1DeleteOptions = new V1DeleteOptions();
-
-                v1DeleteOptions.setApiVersion(podToBeDeleted.getApiVersion());
-                
-                V1Pod v1Status = coreV1Api.deleteNamespacedPod(
-                    podMetadata.getName(),
                     namespace,
                     null,
                     null,
                     null,
                     null,
+                    label.isBlank() ? null : label,
                     null,
-                    v1DeleteOptions
-                );
+                    null,
+                    null,
+                    null,
+                    null);
 
-                log.info("Status: {}", v1Status.getStatus());
+            // Getting a random pod that will be deleted
+            if (!podList.getItems().isEmpty()) {
+                int randomListElementIndex = random.nextInt(podList.getItems().size());
+
+                V1Pod podToBeDeleted = podList.getItems().get(randomListElementIndex);
+
+                V1ObjectMeta podMetadata = podToBeDeleted.getMetadata();
+
+                // Getting pod name
+                if (podMetadata != null) {
+                    String podName = podMetadata.getName();
+                    log.info("The pod {} will be deleted!", podName);
+
+                    V1DeleteOptions v1DeleteOptions = new V1DeleteOptions();
+                    v1DeleteOptions.setApiVersion(podToBeDeleted.getApiVersion());
+
+                    // Pod deletion
+                    V1Pod v1Status = coreV1Api.deleteNamespacedPod(
+                            podName,
+                            namespace,
+                            null,
+                            null,
+                            null,
+                            null,
+                            null,
+                            v1DeleteOptions);
+
+                    log.info("Pod {} deleted! Status: {}", podName, v1Status.getStatus());
+                } else {
+                    log.warn("Pod Metadata is null!");
+                }
             } else {
-                log.warn("Pod Metadata is null!");
+                log.warn("No pods found in namespace {}!", namespace);
             }
-
         } catch (IOException e) {
+
             log.error("Error with ApiClient!", e);
             e.printStackTrace();
+
         } catch (ApiException e) {
+
             log.error("Error in get pod from namespace {} ", namespace);
             log.info(e.getResponseBody());
-            //e.printStackTrace();
+            e.printStackTrace();
+
         }
     }
 }
